Fix invalid TinyDB request.
TinyDB get function expects a query, not a dictionary.
Also, if we use get when multiple items matches, it will return a random item (https://tinydb.readthedocs.io/en/latest/usage.html#retrieving-data).

If we use the all function, then we get an IndexError if we try to select from a blank list.

We can select a specific EID with the get function.  If TinyDB assigns EIDs in order, this should work.